% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{gather}
\alias{gather}
\title{Gathers multiple HTML files into single website}
\usage{
gather(
  files = NULL,
  output_dir = NULL,
  gatherrhtmls_file = "_gatherrhtmls.yaml"
)
}
\arguments{
\item{files}{Character vector of files to be included in the output.
Overwrites 'files' field in gatherrhtmls_file.}

\item{output_dir}{Character, path where \code{gather} will save output files.}

\item{gatherrhtmls_file}{Character, file name with gatherrhtmls configuration
in YAML format. Default: _gatherrhtmls.yaml}
}
\description{
Gathers multiple HTML files into single website. Searches for <title> and
<h1> tags in html file and uses them to create navigation menu. Created menu
and link to CSS stylesheet file are inserted into all given files at the end
of <head> section.
}
\details{
It is recommended to create configuration file using \code{generate_yaml}
and then use \code{gather()} without any further parameters.
}
\examples{
# Use configuration file or default values if "_gatherrhtmls.yaml" doesn't
# exist
gather()

# Build site using two given files and "final" directory
gather(files = c("raport1.html", "analysis2.html"), output_dir = "final")

}
