% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherrhtmls.R
\docType{package}
\name{gatherrhtmls}
\alias{gatherrhtmls}
\title{gatherrhtmls: A package to pick up multiple HTML reports into a single
website.}
\description{
gatherrhtmls: A package to pick up multiple HTML reports into a single
website.
}
\section{Functions}{


\code{gather()}

\code{generate_yaml()}

\code{get_htmls()}
}

