% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_yaml.R
\name{generate_yaml}
\alias{generate_yaml}
\title{Generates configuration file for \code{gather()}}
\usage{
generate_yaml(files = get_htmls(), output_dir = "report")
}
\arguments{
\item{files}{Character vector, of HTML files to be included in configuration.
file}

\item{output_dir}{Character, path where \code{gather} will save output files.}
}
\description{
Generates configuration file for \code{gather()}
}
