% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_htmls.R
\name{get_htmls}
\alias{get_htmls}
\title{Lists HTML files in given path}
\usage{
get_htmls(path = ".")
}
\arguments{
\item{path}{Character. Default: "."}
}
\value{
Character vector of found files.
}
\description{
Lists HTML files in given path
}
